requirements_20240326.txt
adal==1.2.7
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
anyio==4.3.0
applicationinsights==0.11.10
argcomplete==3.1.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==22.2.0
azure-appconfiguration==1.1.1
azure-batch==14.1.0
azure-cli==2.58.0
azure-cli-core==2.58.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.30.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-identity==1.15.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.2.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==30.4.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==29.1.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.4.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b15
azure-mgmt-recoveryservices==2.5.0
azure-mgmt-recoveryservicesbackup==8.0.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.3.0
azure-mgmt-resource==23.1.0b2
azure-mgmt-search==9.1.0
azure-mgmt-security==5.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-servicelinker==1.2.0b1
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b15
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-query==1.2.0
azure-multiapi-storage==1.2.0
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.18.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
bcrypt==4.1.2
billiard==4.2.0
build==1.0.3
CacheControl==0.14.0
cachetools==5.3.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.4
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring-parser==0.15
dulwich==0.21.7
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fabric==3.2.2
fastapi==0.109.2
fastapi-auth0==0.5.0
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
inboard==0.37.0
installer==0.7.0
instructor==0.6.1
invoke==2.2.0
isodate==0.6.1
jaraco.classes==3.3.1
javaproperties==0.5.2
jmespath==1.0.1
jsondiff==2.0.0
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.1
knack==0.11.0
kombu==5.3.5
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-text-splitters==0.0.1
langsmith==0.1.19
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
msal==1.26.0
msal-extensions==1.0.0
msgpack==1.0.7
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
orjson==3.9.15
packaging==23.2
paramiko==3.4.0
passlib==1.7.4
pexpect==4.9.0
pinecone-client==3.1.0
pkginfo==1.9.6
platformdirs==4.2.0
poetry==1.8.1
poetry-core==1.9.0
poetry-plugin-export==1.6.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycomposefile==0.0.30
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0.1
rapidfuzz==3.6.1
raven==6.10.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.0
rsa==4.9
scp==0.13.6
semver==2.13.0
shellingham==1.5.4
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.1
sniffio==1.3.0
SQLAlchemy==2.0.27
sshtunnel==0.1.5
starlette==0.36.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
trove-classifiers==2024.2.23
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.17.6
uvloop==0.19.0
vine==5.1.0
virtualenv==20.25.1
watchgod==0.7
wcwidth==0.2.13
websocket-client==1.3.3
websockets==12.0
wrapt==1.16.0
xattr==1.1.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.17.0
